requirements.txt
flask==2.3.3
flask-cors==4.0.0
python-dotenv==1.0.0
Pillow==10.1.0
PyMuPDF==1.23.8
python-magic==0.4.27
